+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - add: workshop collection creation
 - add: auto set the name of the workshop addon after download
+- add: auto redownload items (disabled by default)
 # v0.4.1
 - improve: select all text when edit text
 - fix: some workshop links can't be parsed